code reorg, tidy, summary, etable method for GelbachD
